# improve textlogger