**What this PR does / why we need it**:

This is a combination of several performance improvements. 

Eventually the goal is to use just that textlogger in Kubernetes, without going through any of the klog APIs like klog.Info nor through the legacy output handling.

**Which issue(s) this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close the issue(s) when PR gets merged)*:
Fixes #360
Fixes #334

**Special notes for your reviewer**:

**Release note**:
```release-note
improved performance
```