@harshanarayana 2023-01-16
> Looks like the vmodule tests are failing again ?

I am running tests locally and not seem to be reproducing. Am I missing something @pohly  ?

```bash
❯ go version                                                                                                                                                                                                                                                                           100%  ▓▒░
go version go1.18 darwin/amd64

..... truncated logs
    --- PASS: TestKlogrInternalStackZapr/log_with_name_and_values (0.00s)
    --- PASS: TestKlogrInternalStackZapr/override_WithValues (0.00s)
    --- PASS: TestKlogrInternalStackZapr/handle_odd-numbers_of_KVs_in_both_log_values_and_Info_args (0.00s)
PASS
ok  	example/output_test	0.857s
?   	example/set_output	[no test files]
?   	example/structured_logging	[no test files]

```