@pohly 2023-01-16
> > Looks like the vmodule tests are failing again ?
> 
> I am running tests locally and not seem to be reproducing.

I am not sure why you couldn't reproduce. It failed inside the `example` directory for me reliably. The root cause was the conceptual problem with `V` assuming to be called directly by the code which logs and the two klogr implementations not doing that.

I initially wanted to disable the failing unit tests, but then would have had to document the bug. At that point I decided that fixing it is easier :grin: It needs a new `VDepth` function, though.