@pohly 2023-01-16
@harshanarayana: tests are passing, except for the known (and IMHO acceptable) API changes:
- klogr: klogger.Enabled/Error/Info can be ignored because the "removed" value receivers are not needed by anyone
- k8s.io/klog/v2/test.InitKLog: this was explicitly marked as "experimental" and still is