@serathius 2023-01-17
Recommend rerunning benchmarks with increased number of tries, but let's not block on them.

/lgtm
/hold

Leaving on hold if you want to benchmark before merge.