@pohly 2023-01-18
I compared the baseline (= the "use output test cases also for benchmarking" commit) with the final state of the PR, using `-count=10`.

My recent vmodule fixes (https://github.com/kubernetes/klog/pull/362#issuecomment-1383841142) unintentionally slowed down the `printWithKLog` function and thus affected benchmark results. I've fixed that before running these benchmarks.

Here's the result for the textlogger:

<details>

```
name                                                                            old time/op    new time/op    delta
TextloggerOutput/override_single_value-36                                         4.60µs ± 7%    2.94µs ±10%  -36.14%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_nil_entry-36                                           5.61µs ± 6%    2.67µs ± 5%  -52.36%  (p=0.000 n=10+10)
TextloggerOutput/Error()_for_nil-36                                               7.94µs ± 6%    6.51µs ± 3%  -18.05%  (p=0.000 n=10+10)
TextloggerOutput/String()_for_nil-36                                              9.31µs ± 8%    6.55µs ±11%  -29.63%  (p=0.000 n=10+10)
TextloggerOutput/struct_keys-36                                                   5.34µs ± 9%    3.83µs ± 5%  -28.36%  (p=0.000 n=9+10)
TextloggerOutput/verbosity_disabled-36                                             162ns ± 2%     167ns ± 3%   +3.34%  (p=0.000 n=10+10)
TextloggerOutput/call_depth-36                                                    4.58µs ± 9%    3.32µs ± 6%  -27.46%  (p=0.000 n=10+10)
TextloggerOutput/print_duplicate_keys_in_arguments-36                             4.40µs ± 8%    3.06µs ± 9%  -30.49%  (p=0.000 n=10+9)
TextloggerOutput/KObjSlice_nil_arg-36                                             3.56µs ± 7%    2.67µs ± 7%  -24.88%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_that_panics-36                                      5.11µs ± 6%    3.82µs ± 3%  -25.19%  (p=0.000 n=10+9)
TextloggerOutput/vmodule-36                                                        668ns ± 7%     657ns ± 6%     ~     (p=0.218 n=10+10)
TextloggerOutput/override_WithValues-36                                           5.37µs ± 8%    3.52µs ± 7%  -34.46%  (p=0.000 n=10+10)
TextloggerOutput/KObjs-36                                                         5.68µs ± 7%    3.75µs ±12%  -33.87%  (p=0.000 n=9+10)
TextloggerOutput/odd_WithValues-36                                                14.6µs ± 7%    10.1µs ± 8%  -30.87%  (p=0.000 n=10+10)
TextloggerOutput/quotation-36                                                     4.33µs ± 7%    3.02µs ±11%  -30.24%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_int_arg-36                                             5.30µs ± 6%    2.80µs ± 9%  -47.15%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_as_value-36                                  4.14µs ± 5%    2.80µs ± 7%  -32.28%  (p=0.000 n=10+10)
TextloggerOutput/handle_integer_keys-36                                           7.02µs ± 7%    4.26µs ± 7%  -39.39%  (p=0.000 n=10+10)
TextloggerOutput/verbosity_enabled-36                                             3.49µs ± 6%    2.65µs ± 8%  -24.01%  (p=0.000 n=10+10)
TextloggerOutput/log_with_values-36                                               4.09µs ± 5%    2.81µs ± 8%  -31.36%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs_in_both_log_values_and_Info_args-36    6.57µs ± 6%    4.04µs ± 6%  -38.45%  (p=0.000 n=10+10)
TextloggerOutput/KObj-36                                                          4.78µs ± 6%    2.60µs ±16%  -45.62%  (p=0.000 n=10+10)
TextloggerOutput/Error()_that_panics-36                                           4.90µs ± 5%    3.40µs ± 8%  -30.58%  (p=0.000 n=10+10)
TextloggerOutput/String()_that_panics-36                                          5.02µs ± 6%    3.35µs ± 6%  -33.35%  (p=0.000 n=10+10)
TextloggerOutput/other_vmodule-36                                                  666ns ± 4%     649ns ± 7%     ~     (p=0.345 n=10+9)
TextloggerOutput/log_with_name_and_values-36                                      4.58µs ± 8%    3.17µs ± 6%  -30.67%  (p=0.000 n=10+9)
TextloggerOutput/log_with_multiple_names_and_values-36                            5.05µs ± 6%    3.82µs ±10%  -24.36%  (p=0.000 n=10+10)
TextloggerOutput/multiple_WithValues-36                                           18.5µs ± 9%    12.5µs ± 6%  -32.48%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_okay-36                                                6.71µs ± 7%    2.73µs ± 8%  -59.33%  (p=0.000 n=10+10)
TextloggerOutput/ignore_MarshalJSON-36                                            4.22µs ±10%    2.78µs ±12%  -34.04%  (p=0.000 n=10+10)
TextloggerOutput/empty_WithValues-36                                              3.44µs ± 6%    2.59µs ± 6%  -24.52%  (p=0.000 n=10+10)
TextloggerOutput/preserve_order_of_key/value_pairs-36                             6.93µs ± 9%    4.17µs ± 8%  -39.87%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_when_using_logr.Error-36                     3.98µs ± 8%    2.57µs ± 7%  -35.34%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_for_nil-36                                          8.27µs ± 7%    3.82µs ± 8%  -53.83%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs-36                                     4.95µs ± 6%    3.29µs ± 7%  -33.53%  (p=0.000 n=9+10)
TextloggerOutput/html_characters-36                                               3.61µs ± 7%    2.81µs ± 8%  -22.11%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_ints-36                                                6.72µs ± 4%    2.81µs ±10%  -58.16%  (p=0.000 n=9+10)
TextloggerOutput/MarshalLog()_that_returns_itself-36                              4.46µs ±10%    2.90µs ± 7%  -34.96%  (p=0.000 n=10+10)
TextloggerOutput/map_keys-36                                                      5.27µs ± 8%    3.85µs ± 7%  -26.86%  (p=0.000 n=10+10)

name                                                                            old alloc/op   new alloc/op   delta
TextloggerOutput/override_single_value-36                                           704B ± 0%      336B ± 0%  -52.27%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_nil_entry-36                                             817B ± 0%      288B ± 0%  -64.75%  (p=0.000 n=10+10)
TextloggerOutput/Error()_for_nil-36                                               1.03kB ± 0%    0.71kB ± 0%  -30.88%  (p=0.000 n=10+10)
TextloggerOutput/String()_for_nil-36                                              1.28kB ± 0%    0.72kB ± 0%  -43.64%  (p=0.000 n=10+10)
TextloggerOutput/struct_keys-36                                                     809B ± 0%      409B ± 0%  -49.44%  (p=0.000 n=10+10)
TextloggerOutput/verbosity_disabled-36                                             64.0B ± 0%     64.0B ± 0%     ~     (all equal)
TextloggerOutput/call_depth-36                                                      680B ± 0%      376B ± 0%  -44.71%  (p=0.000 n=10+10)
TextloggerOutput/print_duplicate_keys_in_arguments-36                               624B ± 0%      320B ± 0%  -48.72%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_nil_arg-36                                               480B ± 0%      288B ± 0%  -40.00%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_that_panics-36                                        705B ± 0%      401B ± 0%  -43.12%  (p=0.000 n=10+10)
TextloggerOutput/vmodule-36                                                        64.0B ± 0%     64.0B ± 0%     ~     (all equal)
TextloggerOutput/override_WithValues-36                                             816B ± 0%      384B ± 0%  -52.94%  (p=0.000 n=10+10)
TextloggerOutput/KObjs-36                                                           865B ± 0%      497B ± 0%  -42.54%  (p=0.000 n=10+10)
TextloggerOutput/odd_WithValues-36                                                2.13kB ± 0%    1.09kB ± 0%  -48.83%  (p=0.000 n=10+10)
TextloggerOutput/quotation-36                                                       624B ± 0%      320B ± 0%  -48.72%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_int_arg-36                                               817B ± 0%      288B ± 0%  -64.75%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_as_value-36                                    608B ± 0%      304B ± 0%  -50.00%  (p=0.000 n=10+10)
TextloggerOutput/handle_integer_keys-36                                           1.15kB ± 0%    0.46kB ± 0%  -60.09%  (p=0.000 n=10+10)
TextloggerOutput/verbosity_enabled-36                                               472B ± 0%      296B ± 0%  -37.29%  (p=0.000 n=10+10)
TextloggerOutput/log_with_values-36                                                 608B ± 0%      304B ± 0%  -50.00%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs_in_both_log_values_and_Info_args-36    1.11kB ± 0%    0.42kB ± 0%  -61.87%  (p=0.000 n=10+10)
TextloggerOutput/KObj-36                                                            680B ± 0%      288B ± 0%  -57.67%  (p=0.000 n=10+10)
TextloggerOutput/Error()_that_panics-36                                             680B ± 0%      377B ± 0%  -44.56%  (p=0.000 n=9+10)
TextloggerOutput/String()_that_panics-36                                            688B ± 0%      385B ± 0%  -44.08%  (p=0.000 n=10+10)
TextloggerOutput/other_vmodule-36                                                  64.0B ± 0%     64.0B ± 0%     ~     (all equal)
TextloggerOutput/log_with_name_and_values-36                                        680B ± 0%      376B ± 0%  -44.71%  (p=0.000 n=10+10)
TextloggerOutput/log_with_multiple_names_and_values-36                              784B ± 0%      480B ± 0%  -38.78%  (p=0.000 n=10+10)
TextloggerOutput/multiple_WithValues-36                                           2.63kB ± 0%    1.48kB ± 0%  -43.81%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_okay-36                                                1.03kB ± 0%    0.29kB ± 0%  -72.12%  (p=0.000 n=10+10)
TextloggerOutput/ignore_MarshalJSON-36                                              608B ± 0%      304B ± 0%  -50.00%  (p=0.000 n=10+10)
TextloggerOutput/empty_WithValues-36                                                464B ± 0%      288B ± 0%  -37.93%  (p=0.000 n=10+10)
TextloggerOutput/preserve_order_of_key/value_pairs-36                             1.18kB ± 0%    0.46kB ± 0%  -60.81%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_when_using_logr.Error-36                       608B ± 0%      304B ± 0%  -50.00%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_for_nil-36                                          1.39kB ± 0%    0.42kB ± 0%  -70.09%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs-36                                       704B ± 0%      336B ± 0%  -52.27%  (p=0.000 n=10+10)
TextloggerOutput/html_characters-36                                                 480B ± 0%      304B ± 0%  -36.67%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_ints-36                                                1.03kB ± 0%    0.30kB ± 0%  -71.35%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_that_returns_itself-36                                624B ± 0%      320B ± 0%  -48.72%  (p=0.000 n=10+10)
TextloggerOutput/map_keys-36                                                        745B ± 0%      441B ± 0%  -40.81%  (p=0.000 n=10+10)

name                                                                            old allocs/op  new allocs/op  delta
TextloggerOutput/override_single_value-36                                           10.0 ± 0%       6.0 ± 0%  -40.00%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_nil_entry-36                                             15.0 ± 0%       4.0 ± 0%  -73.33%  (p=0.000 n=10+10)
TextloggerOutput/Error()_for_nil-36                                                 12.0 ± 0%       9.0 ± 0%  -25.00%  (p=0.000 n=10+10)
TextloggerOutput/String()_for_nil-36                                                13.0 ± 0%       9.0 ± 0%  -30.77%  (p=0.000 n=10+10)
TextloggerOutput/struct_keys-36                                                     14.0 ± 0%      10.0 ± 0%  -28.57%  (p=0.000 n=10+10)
TextloggerOutput/verbosity_disabled-36                                              1.00 ± 0%      1.00 ± 0%     ~     (all equal)
TextloggerOutput/call_depth-36                                                      9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/print_duplicate_keys_in_arguments-36                               9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_nil_arg-36                                               9.00 ± 0%      4.00 ± 0%  -55.56%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_that_panics-36                                        10.0 ± 0%       7.0 ± 0%  -30.00%  (p=0.000 n=10+10)
TextloggerOutput/vmodule-36                                                         1.00 ± 0%      1.00 ± 0%     ~     (all equal)
TextloggerOutput/override_WithValues-36                                             15.0 ± 0%      11.0 ± 0%  -26.67%  (p=0.000 n=10+10)
TextloggerOutput/KObjs-36                                                           17.0 ± 0%      10.0 ± 0%  -41.18%  (p=0.000 n=10+10)
TextloggerOutput/odd_WithValues-36                                                  30.0 ± 0%      18.0 ± 0%  -40.00%  (p=0.000 n=10+10)
TextloggerOutput/quotation-36                                                       8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_int_arg-36                                               12.0 ± 0%       4.0 ± 0%  -66.67%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_as_value-36                                    8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/handle_integer_keys-36                                             16.0 ± 0%      11.0 ± 0%  -31.25%  (p=0.000 n=10+10)
TextloggerOutput/verbosity_enabled-36                                               6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/log_with_values-36                                                 8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs_in_both_log_values_and_Info_args-36      14.0 ± 0%       9.0 ± 0%  -35.71%  (p=0.000 n=10+10)
TextloggerOutput/KObj-36                                                            11.0 ± 0%       4.0 ± 0%  -63.64%  (p=0.000 n=10+10)
TextloggerOutput/Error()_that_panics-36                                             9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/String()_that_panics-36                                            9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/other_vmodule-36                                                   1.00 ± 0%      1.00 ± 0%     ~     (all equal)
TextloggerOutput/log_with_name_and_values-36                                        9.00 ± 0%      6.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/log_with_multiple_names_and_values-36                              12.0 ± 0%       9.0 ± 0%  -25.00%  (p=0.000 n=10+10)
TextloggerOutput/multiple_WithValues-36                                             34.0 ± 0%      20.0 ± 0%  -41.18%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_okay-36                                                  21.0 ± 0%       4.0 ± 0%  -80.95%  (p=0.000 n=10+10)
TextloggerOutput/ignore_MarshalJSON-36                                              8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/empty_WithValues-36                                                6.00 ± 0%      4.00 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/preserve_order_of_key/value_pairs-36                               15.0 ± 0%      10.0 ± 0%  -33.33%  (p=0.000 n=10+10)
TextloggerOutput/regular_error_types_when_using_logr.Error-36                       8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_for_nil-36                                            14.0 ± 0%       6.0 ± 0%  -57.14%  (p=0.000 n=10+10)
TextloggerOutput/handle_odd-numbers_of_KVs-36                                       10.0 ± 0%       6.0 ± 0%  -40.00%  (p=0.000 n=10+10)
TextloggerOutput/html_characters-36                                                 8.00 ± 0%      6.00 ± 0%  -25.00%  (p=0.000 n=10+10)
TextloggerOutput/KObjSlice_ints-36                                                  14.0 ± 0%       5.0 ± 0%  -64.29%  (p=0.000 n=10+10)
TextloggerOutput/MarshalLog()_that_returns_itself-36                                8.00 ± 0%      5.00 ± 0%  -37.50%  (p=0.000 n=10+10)
TextloggerOutput/map_keys-36                                                        14.0 ± 0%      11.0 ± 0%  -21.43%  (p=0.000 n=10+10)

```

</details>

Here are the results for klog itself:

<details>

```
name                                                                      old time/op    new time/op    delta
Header-36                                                                   1.65µs ± 8%    1.65µs ± 7%     ~     (p=0.782 n=10+10)
HeaderWithDir-36                                                            1.71µs ±12%    1.69µs ± 6%     ~     (p=0.494 n=10+10)
V-36                                                                        2.37ns ± 0%    2.37ns ± 0%     ~     (p=0.257 n=10+8)
KRef-36                                                                     0.49ns ± 0%    0.49ns ± 1%   +0.39%  (p=0.027 n=10+8)
KObj-36                                                                     6.42ns ± 0%    6.42ns ± 0%     ~     (p=0.197 n=10+10)
KObjs/0/simple-36                                                            260ns ± 3%     266ns ± 3%     ~     (p=0.055 n=10+10)
KObjs/0/conditional-36                                                      2.61ns ± 1%    2.61ns ± 0%     ~     (p=0.481 n=10+9)
KObjs/10/simple-36                                                          1.20µs ± 4%    1.23µs ± 4%     ~     (p=0.101 n=10+10)
KObjs/10/conditional-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.285 n=10+10)
KObjs/20/simple-36                                                          1.97µs ± 5%    1.99µs ± 4%     ~     (p=0.363 n=10+10)
KObjs/20/conditional-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.445 n=10+10)
KObjs/30/simple-36                                                          2.71µs ± 3%    2.77µs ± 3%   +2.13%  (p=0.015 n=10+10)
KObjs/30/conditional-36                                                     2.60ns ± 0%    2.61ns ± 0%     ~     (p=0.148 n=9+10)
KObjs/40/simple-36                                                          3.47µs ± 3%    3.53µs ± 3%   +1.65%  (p=0.033 n=9+10)
KObjs/40/conditional-36                                                     2.60ns ± 0%    2.61ns ± 0%     ~     (p=0.409 n=9+10)
KObjs/50/simple-36                                                          4.19µs ± 2%    4.25µs ± 2%   +1.42%  (p=0.012 n=10+10)
KObjs/50/conditional-36                                                     2.61ns ± 0%    2.60ns ± 0%   -0.29%  (p=0.000 n=10+10)
KObjs/60/simple-36                                                          5.28µs ± 3%    5.34µs ± 3%     ~     (p=0.218 n=10+10)
KObjs/60/conditional-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.533 n=10+9)
KObjs/70/simple-36                                                          5.92µs ± 2%    6.03µs ± 2%   +1.91%  (p=0.002 n=10+10)
KObjs/70/conditional-36                                                     2.61ns ± 0%    2.61ns ± 0%     ~     (p=0.794 n=9+10)
KObjs/80/simple-36                                                          6.94µs ± 2%    7.01µs ± 3%     ~     (p=0.190 n=10+10)
…